feat(cart): move discount box component and styles into own sub directory
