+++ packages/osc-ecommerce/app/components/Cart/DiscountBox/discount-box.scss
@@ -1,5 +1,5 @@
-@import "../../../../osc-ui/src/styles/settings";
-@import "../../../../osc-ui/src/styles/tools";
+@import "../../../../../osc-ui/src/styles/settings";
+@import "../../../../../osc-ui/src/styles/tools";
 
 @layer components {
     $discount-box-ttl-size: rem($base-fs);
