+++ packages/osc-ecommerce/app/routes/cart.tsx
@@ -17,8 +17,8 @@ import lineItemStyles from 'osc-ui/dist/src-components-LineItem-line-item.css';
 import selectStyles from 'osc-ui/dist/src-components-Select-select.css';
 import textInputStyles from 'osc-ui/dist/src-components-TextInput-text-input.css';
 import invariant from 'tiny-invariant';
+import discountBoxStyles from '~/components/Cart/DiscountBox/discount-box.css';
 import { CartLayout } from '~/components/Cart/Layout';
-import discountBoxStyles from '~/components/Cart/discount-box.css';
 import { getSettingsData } from '~/models/sanity.server';
 import { CART_QUERY } from '~/queries/sanity/cart';
 import type { CartActions } from '~/types/shopify';
